// gabinet-kosmetyczny-frontend-develop/src/app/feature/account/model/worker-account-information.interface.ts
export interface IWorkerAccountInfromation {
    uuid?: string;
    accountUuid?: string;
    name?: string;
    email?: string;
    surname?: string;
    street?: string;
    code?: string;
    town?: string;
    birthday?: Date;
    phone?: string;
    pesel?: string;
    accountType?: string;
    dateOfEmployment?: Date;
    certificates?: string;
}